Refactor stats into shared module
