+++ muckrock/jurisdiction/views.py
@@ -19,6 +19,7 @@ from rest_framework import viewsets
 
 # MuckRock
 from muckrock.agency.models import Agency
+from muckrock.core.stats import collect_stats
 from muckrock.core.views import (
     MRAutocompleteView,
     MRFilterListView,
@@ -114,21 +115,6 @@ class JurisdictionEmbedView(JurisdictionExploreView):
     """View for embedding the map interactive"""
 
     template_name = "jurisdiction/embed.html"
-
-
-def collect_stats(obj, context):
-    """Helper for collecting stats"""
-    statuses = ("rejected", "ack", "processed", "fix", "no_docs", "done", "appealing")
-    requests = obj.get_requests()
-    status_counts = (
-        requests.filter(status__in=statuses)
-        .order_by("status")
-        .values_list("status")
-        .annotate(Count("status"))
-    )
-    context.update({"num_%s" % s: c for s, c in status_counts})
-    context["num_overdue"] = requests.get_overdue().count()
-    context["num_submitted"] = requests.count()
 
 
 def detail(request, fed_slug, state_slug=None, local_slug=None, preview_text=None):
